pyproject.toml
[tool.poetry]
name = "server-backup-assistant"
version = "0.1.0"
description = ""
authors = ["Roman Slyusar"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.8.10"
pyyaml = "^6.0"
schedule = "^1.1.0"
requests = "^2.28.2"
psutil = "^5.9.4"
sqlalchemy = "^2.0.10"
pillow = "^9.5.0"
pystray = "^0.19.4"
customtkinter = "^5.1.3"
python-dotenv = "^1.0.0"
chardet = "^5.1.0"
pywin32 = "^306"
pyinstaller-versionfile = "^2.1.1"
tee = "0.0.3"
decorator = "^5.1.1"
pyuac = "^0.0.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
pyinstaller = "^5.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
